Rules and Description
The rules are as follows:

- if an alive cell has either 2 or 3 neighbouring alive cells, it will stay alive.
- if an alive cell has less than 2 neighbouring alive cells it will die.
- if an alive cell has more than 3 neighbouring alive cells it will die.
- if a dead cell has exactly 3 alive neighbouring cells it will live.

The grid is randomised to begin, then changes based off of the rules each generation, in order to move through generations you must press n on the keyboard. My grid is modelled after old phone games, such as snake where the board doesnt ended it simply extends to the other side. If you go out the left side you will come in the right side. Neighbouring cells can be on adjacent sides of the grid. cells[0][0] is neighbouring to cells[0][size-1] for example.